Added new unit
Added sprites and class for a new archer unit. Made some changes to UnitCreationPanel and the i18n classes to reflect this addition.
Fixed a bug in Unit that made it skip frames in animations (it thought the animation was over after the first frame).
Changed username in project properties to match github account name.