MenuPanel
Added a MenuPanel subpanel to the MainFrame.
Created localization functions for all the gui elements to be called whenever the language is changed.
The ActionListener from the MainFrame is now passed to its children panels os that all of them trigger the same function in the top-level frame when a button is pressed.
Changed a couple of images.